+++ lib/l10n/app_localizations.dart
@@ -776,6 +776,12 @@ abstract class AppLocalizations {
   /// **'Close'**
   String get close;
 
+  /// No description provided for @dismiss.
+  ///
+  /// In en, this message translates to:
+  /// **'Dismiss'**
+  String get dismiss;
+
   /// No description provided for @savedToDownloads.
   ///
   /// In en, this message translates to: